finished data parsing & formatting
