--- src/components/calendarTimeslots.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-// import styled from "styled-components";
-
-export default function CalendarTimeSlot() {
-  return <div>calendar</div>;
-}
